🧹 Split validate_remediation_commands.py into a validators package
The single file had grown to 2,500 lines holding seven validators. Move
each validator family into its own module under
content/validation/validators/ (aws, azure, oci, gcloud, digitalocean,
openapi, nutanix), with shared helpers (bash-block extraction, command
splitting, failure collection, GitHub annotations) in common.py and the
CLI dispatch in main.py.

validate_remediation_commands.py stays as the entry point, now a thin
shim, so CI invocations and docs are unchanged. Code is moved verbatim
except for one fix: the per-validator
`relative_to(Path.cwd())` annotation paths crashed when the script ran
from any directory other than the repo root; they are replaced by a
shared policy_relpath() helper that is cwd-independent.

All seven validators produce identical results after the split
(aws 837, azure 345, gcp 361, oci 94, digitalocean 43, cloudflare 29,
nutanix 23 — all passing, exit 0).
