+++ content/validation/validate_remediation_commands.py
@@ -15,2518 +15,7 @@
 #   python3 validate_remediation_commands.py cloudflare    # validate Cloudflare API curl commands only
 #   python3 validate_remediation_commands.py nutanix       # validate Nutanix ncli commands only
 
-import concurrent.futures
-import json
-import re
-import shlex
-import subprocess
-import sys
-import urllib.request
-from collections.abc import Iterator
-from pathlib import Path
-
-SCRIPT_DIR = Path(__file__).parent
-CMD_DATA_DIR = SCRIPT_DIR / "cmd_data"
-
-VALIDATORS = ["aws", "azure", "oci", "gcp", "digitalocean", "cloudflare", "nutanix"]
-
-# Collected failures for annotation output.  Each entry is a dict with keys:
-# file, line, uid, command, errors, cloud
-FAILURES: list[dict] = []
-
-
-# ---------------------------------------------------------------------------
-# Shared helpers
-# ---------------------------------------------------------------------------
-
-def extract_bash_blocks(content: str) -> list[tuple[str, int, str]]:
-    """Extract bash code blocks from cli remediation sections.
-
-    Returns a list of (block_text, line_number, uid) tuples where line_number
-    is the 1-based line of the first code line in the block, and uid is the
-    check UID that contains this remediation block.
-    """
-    # Pre-compute a list of (line_number, uid) from all `- uid:` lines so we
-    # can look up the enclosing check for any position in the file.
-    lines = content.split("\n")
-    uid_positions: list[tuple[int, str]] = []
-    for i, line in enumerate(lines):
-        m = re.match(r"^  - uid:\s+(\S+)", line)
-        if m:
-            uid_positions.append((i + 1, m.group(1)))
-
-    def find_uid_for_line(line_num: int) -> str:
-        """Find the nearest uid defined before line_num."""
-        result = ""
-        for pos, uid in uid_positions:
-            if pos <= line_num:
-                result = uid
-            else:
-                break
-        return result
-
-    pattern = re.compile(
-        r"- id: cli\s*\n\s+desc: \|\s*\n(.*?)(?=\n\s+- id: |\n\s+refs:|\n  - uid: |\Z)",
-        re.DOTALL,
-    )
-    blocks = []
-    for match in pattern.finditer(content):
-        desc_block = match.group(1)
-        desc_start = match.start(1)
-        # Line number of the cli remediation block itself
-        cli_line = content[:match.start()].count("\n") + 1
-        uid = find_uid_for_line(cli_line)
-
-        for fence in re.finditer(r"```bash\s*\n(.*?)```", desc_block, re.DOTALL):
-            block = fence.group(1).strip()
-            if block:
-                code_offset = desc_start + fence.start(1)
-                line_number = content[:code_offset].count("\n") + 1
-                blocks.append((block, line_number, uid))
-    return blocks
-
-
-def split_commands(block: str, prefix: str, block_start_line: int) -> list[tuple[str, int]]:
-    """Split a code block into individual commands starting with prefix.
-
-    Returns a list of (command, line_number) tuples.
-    """
-    lines = block.split("\n")
-    commands = []
-    i = 0
-
-    while i < len(lines):
-        line = lines[i]
-        raw_line_num = block_start_line + i
-
-        # Join continuation lines
-        full_line = line
-        cont_lines = 0
-        while full_line.rstrip().endswith("\\") and i + cont_lines + 1 < len(lines):
-            cont_lines += 1
-            full_line = full_line.rstrip()[:-1] + " " + lines[i + cont_lines].strip()
-
-        stripped = full_line.strip()
-        if stripped and not stripped.startswith("#"):
-            # Use shlex to handle quoted values containing | or ;
-            # then re-join and split on unquoted pipes/semicolons
-            try:
-                tokens = shlex.split(stripped)
-            except ValueError:
-                tokens = stripped.split()
-            rejoined = " ".join(tokens)
-            # Split on pipe/semicolon boundaries
-            for segment in re.split(r"\s*[|;]\s*", rejoined):
-                segment = segment.strip()
-                if segment.startswith(f"{prefix} "):
-                    commands.append((segment, raw_line_num))
-
-        i += 1 + cont_lines
-
-    return commands
-
-
-def truncate_cmd(cmd: str, max_len: int = 120) -> str:
-    """Collapse whitespace and truncate a command for display."""
-    display = " ".join(cmd.split())
-    if len(display) > max_len:
-        display = display[: max_len - 3] + "..."
-    return display
-
-
-# ---------------------------------------------------------------------------
-# AWS validation
-# ---------------------------------------------------------------------------
-#
-# The AWS commands database is built in-memory at validation time by
-# introspecting the botocore service models bundled with AWS CLI v2. This
-# avoids shipping a large, frequently-churning aws_commands.json in the repo.
-# Requires the AWS CLI v2 to be installed in the environment running validation.
-
-AWS_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-aws-security.mql.yaml"
-
-# CLI-name → botocore-model-name aliases. Most CLI services match their
-# botocore names; these differ and need an explicit mapping.
-AWS_CLI_TO_BOTOCORE = {
-    "configservice": "config",
-    "s3api": "s3",
-}
-
-# Global flags available on every AWS CLI command.
-AWS_GLOBAL_FLAGS = [
-    "--ca-bundle",
-    "--cli-auto-prompt",
-    "--cli-binary-format",
-    "--cli-connect-timeout",
-    "--cli-error-format",
-    "--cli-input-json",
-    "--cli-input-yaml",
-    "--cli-read-timeout",
-    "--color",
-    "--debug",
-    "--endpoint-url",
-    "--generate-cli-skeleton",
-    "--no-cli-auto-prompt",
-    "--no-cli-pager",
-    "--no-paginate",
-    "--no-sign-request",
-    "--no-verify-ssl",
-    "--output",
-    "--profile",
-    "--query",
-    "--region",
-    "--version",
-]
-
-_XFORM_RE1 = re.compile(r"([a-z0-9])([A-Z])")
-_XFORM_RE2 = re.compile(r"([A-Z])([A-Z][a-z])")
-
-
-def _xform_name(name: str) -> str:
-    """CamelCase → snake_case (mirrors botocore.xform_name)."""
-    s = _XFORM_RE1.sub(r"\1_\2", name)
-    s = _XFORM_RE2.sub(r"\1_\2", s)
-    return s.lower()
-
-
-def _load_aws_service_ops(service_dir: Path) -> tuple[list[str], dict[str, list[str]]]:
-    """Parse `<service>/<api-version>/service-2.json` and return
-    (subcommands, {subcommand: flags}) exactly as the AWS CLI exposes them.
-    """
-    candidates = sorted(service_dir.glob("*/service-2.json"))
-    if not candidates:
-        return [], {}
-    data = json.loads(candidates[-1].read_text())
-    shapes = data.get("shapes", {})
-    subcommands: list[str] = []
-    flag_map: dict[str, list[str]] = {}
-    for op_name, op in data.get("operations", {}).items():
-        subcmd = _xform_name(op_name).replace("_", "-")
-        subcommands.append(subcmd)
-        flags: list[str] = []
-        input_ref = op.get("input")
-        if input_ref:
-            input_shape = shapes.get(input_ref.get("shape", ""), {})
-            for m_name, m_ref in input_shape.get("members", {}).items():
-                kebab = _xform_name(m_name).replace("_", "-")
-                flags.append(f"--{kebab}")
-                m_shape = shapes.get(m_ref.get("shape", ""), {})
-                if m_shape.get("type") == "boolean":
-                    flags.append(f"--no-{kebab}")
-        flag_map[subcmd] = sorted(set(flags))
-    return sorted(subcommands), flag_map
-
-
-def detect_aws_services_from_policy() -> list[str]:
-    """Return the set of top-level aws CLI services used in the AWS policy."""
-    if not AWS_POLICY_FILE.exists():
-        return []
-
-    content = AWS_POLICY_FILE.read_text()
-    services = set()
-    for match in re.finditer(r"```bash\s*\n(.*?)```", content, re.DOTALL):
-        block = match.group(1)
-        joined = re.sub(r"\\\s*\n\s*", " ", block)
-        for line in joined.split("\n"):
-            line = line.strip()
-            if line.startswith("aws "):
-                parts = line.split()
-                if len(parts) >= 2 and not parts[1].startswith("-"):
-                    services.add(parts[1])
-    return sorted(services)
-
-
-def find_aws_botocore_data_dir() -> Path | None:
-    """Locate the `botocore/data/` directory bundled with AWS CLI v2.
-
-    Handles both install layouts:
-      - Homebrew / pip:      `.../lib/python3.X/site-packages/awscli/botocore/data/`
-      - Official installer:  `.../aws-cli/v2/<ver>/dist/awscli/botocore/data/`
-
-    Returns None if AWS CLI v2 is not installed or the data dir is missing.
-    """
-    aws_path = subprocess.run(
-        ["which", "aws"], capture_output=True, text=True
-    ).stdout.strip()
-    if not aws_path:
-        return None
-
-    real_path = Path(aws_path).resolve()
-    # The install root is usually 2–3 ancestors above the aws binary.
-    search_roots = [real_path.parent.parent, real_path.parent.parent.parent]
-    for root in search_roots:
-        for pattern in (
-            "**/dist/awscli/botocore/data",
-            "**/site-packages/awscli/botocore/data",
-        ):
-            candidates = list(root.glob(pattern))
-            if candidates:
-                return candidates[0]
-    return None
-
-
-def build_aws_commands_db() -> dict[str, list[str]]:
-    """Build an in-memory AWS commands database by parsing the botocore
-    service-2.json files bundled with AWS CLI v2, scoped to services used
-    in the AWS policy.
-
-    Returns an empty dict if the policy file has no aws commands. Exits with
-    a helpful error if the AWS CLI v2 is not installed.
-    """
-    services = detect_aws_services_from_policy()
-    if not services:
-        return {}
-
-    data_dir = find_aws_botocore_data_dir()
-    if not data_dir:
-        print(
-            "Error: AWS CLI v2 not found in PATH.\n"
-            "\n"
-            "Validating aws remediation commands requires the AWS CLI v2 to\n"
-            "be installed locally — the botocore service models bundled with\n"
-            "the CLI are the source of truth for valid commands and flags.\n"
-            "The AWS CLI v1 (pip-installed) is not supported.\n"
-            "\n"
-            "Install options:\n"
-            "  macOS (Homebrew): brew install awscli\n"
-            "  macOS (pkg):      https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html#getting-started-install-macos\n"
-            "  Linux:            https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html#cliv2-linux-install\n"
-            "  Windows:          https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html#cliv2-windows-install\n"
-            "\n"
-            "After installing, ensure `aws` is on your PATH and re-run:\n"
-            f"  python3 {sys.argv[0]} aws\n"
-            "\n"
-            "To skip aws validation, run a different target instead\n"
-            "(e.g. `azure`, `oci`, or `gcp`).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    commands: dict[str, list[str]] = {}
-    for cli_name in services:
-        botocore_name = AWS_CLI_TO_BOTOCORE.get(cli_name, cli_name)
-        service_dir = data_dir / botocore_name
-        if not service_dir.is_dir():
-            print(
-                f"Warning: AWS service '{botocore_name}' not found under {data_dir}",
-                file=sys.stderr,
-            )
-            continue
-        subcommands, flag_map = _load_aws_service_ops(service_dir)
-        commands[cli_name] = subcommands
-        for subcmd, flags in flag_map.items():
-            commands[f"{cli_name} {subcmd}"] = sorted(set(flags + AWS_GLOBAL_FLAGS))
-    return commands
-
-
-def parse_aws_command(cmd: str) -> tuple[str, str, list[str]]:
-    """Parse an aws command into (service, subcommand, flags).
-
-    Always returns a non-empty service when the command starts with 'aws'
-    followed by at least one non-flag token, so the caller can report
-    unknown services/subcommands instead of silently skipping them.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 2 or tokens[0] != "aws":
-        return "", "", []
-
-    service = tokens[1]
-    if service.startswith("-"):
-        return "", "", []
-
-    subcommand = ""
-    flags = []
-    for token in tokens[2:]:
-        if token.startswith("--"):
-            flags.append(token.split("=")[0])
-        elif not subcommand and not token.startswith("-"):
-            subcommand = token
-
-    return service, subcommand, flags
-
-
-# The AWS CLI synthesizes these subcommands and flags at runtime; they are
-# real CLI surface but never appear in the botocore service-2.json model the
-# database is built from.
-AWS_CLI_CUSTOM_SUBCOMMANDS = {
-    # Every service with a waiters-2.json model gets an `aws <service> wait`
-    # subcommand whose own arguments are waiter names, not model operations.
-    "wait",
-}
-
-AWS_CLI_CUSTOM_FLAGS = {
-    # CLI customization that writes the seed material to a local file.
-    "iam create-virtual-mfa-device": ["--outfile", "--bootstrap-method"],
-    # Members typed as AttributeBooleanValue structures get boolean-style
-    # --flag/--no-flag forms in the CLI instead of a value argument.
-    "ec2 modify-subnet-attribute": [
-        "--map-public-ip-on-launch",
-        "--no-map-public-ip-on-launch",
-        "--assign-ipv6-address-on-creation",
-        "--no-assign-ipv6-address-on-creation",
-    ],
-}
-
-
-def validate_aws_command(
-    service: str,
-    subcommand: str,
-    flags: list[str],
-    commands_db: dict[str, list[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed AWS command against the commands database."""
-    errors = []
-
-    if service not in commands_db:
-        errors.append(f"unknown service '{service}'")
-        return False, errors
-
-    if not subcommand:
-        errors.append(f"missing subcommand for '{service}'")
-        return False, errors
-
-    if subcommand in AWS_CLI_CUSTOM_SUBCOMMANDS:
-        return True, []
-
-    valid_subcommands = commands_db[service]
-    if subcommand not in valid_subcommands:
-        errors.append(f"unknown subcommand '{service} {subcommand}'")
-        return False, errors
-
-    key = f"{service} {subcommand}"
-    if key in commands_db:
-        valid_flags = set(commands_db[key])
-        valid_flags.update(AWS_CLI_CUSTOM_FLAGS.get(key, []))
-        for flag in flags:
-            if flag not in valid_flags:
-                errors.append(f"unknown flag '{flag}' for '{service} {subcommand}'")
-
-    return len(errors) == 0, errors
-
-
-def validate_aws() -> tuple[int, int]:
-    """Validate AWS CLI commands. Returns (pass_count, fail_count)."""
-    if not AWS_POLICY_FILE.exists():
-        print(f"Error: Policy file not found: {AWS_POLICY_FILE}", file=sys.stderr)
-        sys.exit(1)
-
-    commands_db = build_aws_commands_db()
-    if not commands_db:
-        return 0, 0
-
-    content = AWS_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-
-    pass_count = 0
-    fail_count = 0
-
-    policy_relpath = str(AWS_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "aws", block_line)
-        for cmd, line_num in commands:
-            service, subcommand, flags = parse_aws_command(cmd)
-
-            if not service:
-                continue
-
-            is_valid, errors = validate_aws_command(
-                service, subcommand, flags, commands_db
-            )
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": truncate_cmd(cmd),
-                    "errors": errors,
-                    "cloud": "aws",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# Azure validation
-# ---------------------------------------------------------------------------
-
-AZURE_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-azure-security.mql.yaml"
-M365_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-m365-security.mql.yaml"
-AZURE_COMMANDS_FILE = CMD_DATA_DIR / "azure_commands.json"
-
-# Policy files whose `id: cli` remediations use the Azure CLI (`az`). The M365
-# policy is included here because its CLI remediations also target `az`.
-AZURE_CLI_POLICY_FILES = [AZURE_POLICY_FILE, M365_POLICY_FILE]
-
-
-def parse_az_command(cmd: str, commands_db: dict[str, list[str]]) -> tuple[str, list[str]]:
-    """Parse an az command into (command_path, flags).
-
-    Azure CLI commands have variable-depth paths (e.g. 'network nsg rule create').
-    We match the longest known command path from the database. If no match is
-    found, the raw command path (tokens before the first flag) is returned so
-    the caller can report it as unknown.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 2 or tokens[0] != "az":
-        return "", []
-
-    # Find the longest matching command path by consuming tokens until
-    # we hit a flag or run out of matching commands
-    parts = tokens[1:]  # skip 'az'
-    command_path = ""
-    raw_path_parts = []
-    for i in range(len(parts)):
-        if parts[i].startswith("-"):
-            break
-        raw_path_parts.append(parts[i])
-        candidate = " ".join(parts[: i + 1])
-        if candidate in commands_db:
-            command_path = candidate
-
-    # If no match found, return the raw path so the caller can report it
-    if not command_path and raw_path_parts:
-        command_path = " ".join(raw_path_parts)
-
-    # Extract flags from remaining tokens
-    flags = []
-    for token in parts:
-        if token.startswith("--"):
-            flag = token.split("=")[0]
-            flags.append(flag)
-
-    return command_path, flags
-
-
-def validate_az_command(
-    command_path: str,
-    flags: list[str],
-    commands_db: dict[str, list[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed Azure CLI command against the commands database."""
-    errors = []
-
-    if command_path not in commands_db:
-        errors.append(f"unknown command 'az {command_path}'")
-        return False, errors
-
-    valid_flags = set(commands_db[command_path])
-    for flag in flags:
-        if flag not in valid_flags:
-            errors.append(f"unknown flag '{flag}' for 'az {command_path}'")
-
-    return len(errors) == 0, errors
-
-
-def validate_azure() -> tuple[int, int]:
-    """Validate Azure CLI commands. Returns (pass_count, fail_count).
-
-    Scans both the Azure security policy and the M365 security policy, since
-    M365 `id: cli` remediations also use the Azure CLI (`az`).
-    """
-    if not AZURE_COMMANDS_FILE.exists():
-        print(
-            f"Error: Commands database not found: {AZURE_COMMANDS_FILE}\n"
-            f"Run dump_azure_commands.py first to generate it.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    commands_db = json.loads(AZURE_COMMANDS_FILE.read_text())
-
-    pass_count = 0
-    fail_count = 0
-
-    for policy_file in AZURE_CLI_POLICY_FILES:
-        if not policy_file.exists():
-            print(f"Error: Policy file not found: {policy_file}", file=sys.stderr)
-            sys.exit(1)
-
-        content = policy_file.read_text()
-        blocks = extract_bash_blocks(content)
-        policy_relpath = str(policy_file.resolve().relative_to(Path.cwd()))
-
-        for block_text, block_line, uid in blocks:
-            commands = split_commands(block_text, "az", block_line)
-            for cmd, line_num in commands:
-                command_path, flags = parse_az_command(cmd, commands_db)
-
-                if not command_path:
-                    continue
-
-                is_valid, errors = validate_az_command(
-                    command_path, flags, commands_db
-                )
-
-                if is_valid:
-                    print(f"[PASS] {uid}")
-                    print(f"       {truncate_cmd(cmd)}")
-                    pass_count += 1
-                else:
-                    print(f"[FAIL] {uid}")
-                    print(f"       {truncate_cmd(cmd)}")
-                    for error in errors:
-                        print(f"       {error}")
-                    fail_count += 1
-                    FAILURES.append({
-                        "file": policy_relpath,
-                        "line": line_num,
-                        "uid": uid,
-                        "command": truncate_cmd(cmd),
-                        "errors": errors,
-                        "cloud": "azure",
-                    })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# OCI validation
-# ---------------------------------------------------------------------------
-#
-# The OCI commands database is built in-memory at validation time by walking
-# the Click command tree from the oci_cli Python package. This avoids shipping
-# a large oci_commands.json in the repo. Requires the OCI CLI to be installed
-# in the environment running validation.
-
-OCI_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-oci-security.mql.yaml"
-
-# Global flags available on every OCI CLI command.
-OCI_GLOBAL_FLAGS = [
-    "--auth",
-    "--auth-purpose",
-    "--cert-bundle",
-    "--cli-auto-prompt",
-    "--cli-rc-file",
-    "--config-file",
-    "--connection-timeout",
-    "--debug",
-    "--defaults-file",
-    "--enable-dual-stack",
-    "--enable-propagation",
-    "--endpoint",
-    "--generate-full-command-json-input",
-    "--generate-param-json-input",
-    "--help",
-    "--max-retries",
-    "--no-retry",
-    "--opc-client-request-id",
-    "--opc-request-id",
-    "--output",
-    "--profile",
-    "--proxy",
-    "--query",
-    "--raw-output",
-    "--read-timeout",
-    "--realm-specific-endpoint",
-    "--region",
-    "--request-id",
-]
-
-# Script that runs inside the OCI CLI's bundled Python to extract the command
-# tree from the Click-based CLI.
-_OCI_EXTRACT_SCRIPT = r"""
-import json, sys
-sys.path.insert(0, sys.argv[1])
-
-import click
-from oci_cli import dynamic_loader
-from oci_cli.cli import cli
-
-dynamic_loader.load_all_services()
-
-services = json.loads(sys.argv[2])
-
-def walk(group, prefix=""):
-    results = {}
-    if isinstance(group, click.Group):
-        subcommands = []
-        for name, cmd in sorted(group.commands.items()):
-            path = f"{prefix} {name}".strip()
-            if isinstance(cmd, click.Group):
-                subcommands.append(name)
-                results.update(walk(cmd, path))
-            else:
-                subcommands.append(name)
-                flags = []
-                for param in cmd.params:
-                    for opt in param.opts:
-                        if opt.startswith("--"):
-                            flags.append(opt)
-                    if hasattr(param, "secondary_opts"):
-                        for opt in param.secondary_opts:
-                            if opt.startswith("--"):
-                                flags.append(opt)
-                results[path] = sorted(set(flags))
-        if prefix:
-            results[prefix] = sorted(subcommands)
-    return results
-
-result = {}
-for svc in services:
-    if svc in cli.commands:
-        result.update(walk(cli.commands[svc], svc))
-    else:
-        print(f"Warning: service '{svc}' not found in OCI CLI", file=sys.stderr)
-
-print(json.dumps(result))
-"""
-
-
-def detect_oci_services_from_policy() -> list[str]:
-    """Return the set of top-level oci services used in the OCI policy."""
-    if not OCI_POLICY_FILE.exists():
-        return []
-
-    content = OCI_POLICY_FILE.read_text()
-    services = set()
-    for match in re.finditer(r"```bash\s*\n(.*?)```", content, re.DOTALL):
-        block = match.group(1)
-        joined = re.sub(r"\\\s*\n\s*", " ", block)
-        for line in joined.split("\n"):
-            line = line.strip()
-            if line.startswith("oci "):
-                parts = line.split()
-                if len(parts) >= 2 and not parts[1].startswith("-"):
-                    services.add(parts[1])
-    return sorted(services)
-
-
-def _find_oci_binary() -> str | None:
-    oci_path = subprocess.run(
-        ["which", "oci"], capture_output=True, text=True
-    ).stdout.strip()
-    return oci_path or None
-
-
-def find_oci_cli_site_packages(oci_path: str) -> str | None:
-    """Locate the site-packages directory containing oci_cli."""
-    real_path = Path(oci_path).resolve()
-    base = real_path.parent.parent
-    candidates = list(base.glob("**/site-packages/oci_cli/__init__.py"))
-    if not candidates:
-        return None
-    return str(candidates[0].parent.parent)
-
-
-def find_oci_python(oci_path: str) -> str | None:
-    """Return the Python interpreter used by the OCI CLI."""
-    try:
-        with open(oci_path) as f:
-            first_line = f.readline().strip()
-    except OSError:
-        first_line = ""
-
-    if first_line.startswith("#!"):
-        python_path = first_line[2:].strip()
-        if Path(python_path).exists():
-            return python_path
-
-    real_path = Path(oci_path).resolve()
-    base = real_path.parent.parent
-    candidates = list(base.glob("**/bin/python3"))
-    if candidates:
-        return str(candidates[0])
-    return None
-
-
-def build_oci_commands_db() -> dict[str, list[str]]:
-    """Build an in-memory OCI commands database by walking the oci_cli Click
-    command tree, scoped to services used in the OCI policy.
-
-    Returns an empty dict if the policy file has no oci commands. Exits with
-    a helpful error if the OCI CLI is not installed.
-    """
-    services = detect_oci_services_from_policy()
-    if not services:
-        return {}
-
-    oci_path = _find_oci_binary()
-    if not oci_path:
-        print(
-            "Error: oci CLI not found in PATH.\n"
-            "\n"
-            "Validating oci remediation commands requires the OCI CLI to be\n"
-            "installed locally — the oci_cli Python package's Click command\n"
-            "tree is the source of truth for valid commands and flags.\n"
-            "\n"
-            "Install options:\n"
-            "  pip:              pip install oci-cli\n"
-            "  macOS (Homebrew): brew install oci-cli\n"
-            "  Installer script: https://docs.oracle.com/en-us/iaas/Content/API/SDKDocs/cliinstall.htm\n"
-            "\n"
-            "After installing, ensure `oci` is on your PATH and re-run:\n"
-            f"  python3 {sys.argv[0]} oci\n"
-            "\n"
-            "To skip oci validation, run a different target instead\n"
-            "(e.g. `aws`, `azure`, or `gcp`).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    site_packages = find_oci_cli_site_packages(oci_path)
-    python_path = find_oci_python(oci_path)
-    if not site_packages or not python_path:
-        print(
-            f"Error: Found `oci` at '{oci_path}' but could not locate its\n"
-            f"bundled Python environment (site-packages: {site_packages}, "
-            f"python: {python_path}).\n"
-            "\n"
-            "This usually indicates a broken or partial CLI install. Try\n"
-            "reinstalling the OCI CLI (see install options above).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    result = subprocess.run(
-        [
-            python_path,
-            "-c",
-            _OCI_EXTRACT_SCRIPT,
-            site_packages,
-            json.dumps(services),
-        ],
-        capture_output=True,
-        text=True,
-        timeout=120,
-    )
-
-    if result.returncode != 0:
-        print(
-            f"Error extracting OCI commands from oci_cli:\n{result.stderr}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    if result.stderr:
-        for line in result.stderr.splitlines():
-            if line.strip() and "Warning" not in line:
-                print(line, file=sys.stderr)
-
-    commands = json.loads(result.stdout)
-    for key, value in commands.items():
-        if value and isinstance(value[0], str) and value[0].startswith("-"):
-            commands[key] = sorted(set(value + OCI_GLOBAL_FLAGS))
-    return commands
-
-
-def parse_oci_command(cmd: str, commands_db: dict[str, list[str]]) -> tuple[str, list[str]]:
-    """Parse an oci command into (command_path, flags).
-
-    OCI CLI commands have variable-depth paths (e.g. 'iam user api-key list').
-    We match the longest known command path from the database. If no match is
-    found, the raw command path (tokens before the first flag) is returned so
-    the caller can report it as unknown.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 2 or tokens[0] != "oci":
-        return "", []
-
-    # Find the longest matching command path by consuming tokens until
-    # we hit a flag or run out of matching commands
-    parts = tokens[1:]  # skip 'oci'
-    command_path = ""
-    raw_path_parts = []
-    for i in range(len(parts)):
-        if parts[i].startswith("-"):
-            break
-        raw_path_parts.append(parts[i])
-        candidate = " ".join(parts[: i + 1])
-        if candidate in commands_db:
-            command_path = candidate
-
-    # If no match found, return the raw path so the caller can report it
-    if not command_path and raw_path_parts:
-        command_path = " ".join(raw_path_parts)
-
-    # Extract flags from remaining tokens
-    flags = []
-    for token in parts:
-        if token.startswith("--"):
-            flag = token.split("=")[0]
-            flags.append(flag)
-
-    return command_path, flags
-
-
-def validate_oci_command(
-    command_path: str,
-    flags: list[str],
-    commands_db: dict[str, list[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed OCI CLI command against the commands database."""
-    errors = []
-
-    if command_path not in commands_db:
-        errors.append(f"unknown command 'oci {command_path}'")
-        return False, errors
-
-    valid_flags = set(commands_db[command_path])
-    for flag in flags:
-        if flag not in valid_flags:
-            errors.append(f"unknown flag '{flag}' for 'oci {command_path}'")
-
-    return len(errors) == 0, errors
-
-
-def validate_oci() -> tuple[int, int]:
-    """Validate OCI CLI commands. Returns (pass_count, fail_count)."""
-    if not OCI_POLICY_FILE.exists():
-        print(f"Error: Policy file not found: {OCI_POLICY_FILE}", file=sys.stderr)
-        sys.exit(1)
-
-    commands_db = build_oci_commands_db()
-    if not commands_db:
-        return 0, 0
-
-    content = OCI_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-
-    pass_count = 0
-    fail_count = 0
-
-    policy_relpath = str(OCI_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "oci", block_line)
-        for cmd, line_num in commands:
-            command_path, flags = parse_oci_command(cmd, commands_db)
-
-            if not command_path:
-                continue
-
-            is_valid, errors = validate_oci_command(
-                command_path, flags, commands_db
-            )
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": truncate_cmd(cmd),
-                    "errors": errors,
-                    "cloud": "oci",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# gcloud validation
-# ---------------------------------------------------------------------------
-#
-# Unlike aws/azure/oci, the gcloud commands database is built in-memory at
-# validation time from the Google Cloud SDK's static completion tree. This
-# avoids shipping a large, frequently-churning gcloud_commands.json in the
-# repo. Requires gcloud to be installed in the environment running validation.
-
-GCLOUD_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-gcp-security.mql.yaml"
-
-
-def detect_gcloud_services_from_policy() -> list[str]:
-    """Return the set of top-level gcloud services used in the GCP policy."""
-    if not GCLOUD_POLICY_FILE.exists():
-        return []
-
-    content = GCLOUD_POLICY_FILE.read_text()
-    services = set()
-    for match in re.finditer(r"```bash\s*\n(.*?)```", content, re.DOTALL):
-        block = match.group(1)
-        joined = re.sub(r"\\\s*\n\s*", " ", block)
-        for line in joined.split("\n"):
-            line = line.strip()
-            if line.startswith("gcloud "):
-                parts = line.split()
-                if len(parts) >= 2 and not parts[1].startswith("-"):
-                    services.add(parts[1])
-    return sorted(services)
-
-
-def find_gcloud_completions_path() -> str | None:
-    """Locate the gcloud static completion tree directory, or None if missing."""
-    which = subprocess.run(
-        ["which", "gcloud"], capture_output=True, text=True
-    )
-    gcloud_path = which.stdout.strip()
-    if not gcloud_path:
-        return None
-
-    real_path = Path(gcloud_path).resolve()
-    sdk_root = real_path.parent.parent
-
-    completions = sdk_root / "data" / "cli" / "gcloud_completions.py"
-    if completions.exists():
-        return str(completions.parent)
-
-    candidates = list(sdk_root.glob("**/data/cli/gcloud_completions.py"))
-    if candidates:
-        return str(candidates[0].parent)
-
-    share_candidates = list(
-        Path("/opt/homebrew/share/google-cloud-sdk/data/cli").glob(
-            "gcloud_completions.py"
-        )
-    )
-    if share_candidates:
-        return str(share_candidates[0].parent)
-
-    return None
-
-
-def walk_gcloud_tree(node: dict, prefix: str = "") -> dict[str, list[str]]:
-    """Flatten the gcloud completion tree into {path: [subcommands|flags]}.
-
-    Boolean flags get synthetic --no- variants since gcloud generates these.
-    """
-    results: dict[str, list[str]] = {}
-    commands = node.get("commands", {})
-    if not commands:
-        return results
-
-    subcommand_names = []
-    for name, child in sorted(commands.items()):
-        path = f"{prefix} {name}".strip()
-        child_commands = child.get("commands", {})
-
-        if child_commands:
-            subcommand_names.append(name)
-            results.update(walk_gcloud_tree(child, path))
-        else:
-            subcommand_names.append(name)
-            flags_dict = child.get("flags", {})
-            flags = []
-            for flag_name, flag_type in flags_dict.items():
-                flags.append(flag_name)
-                if flag_type == "bool" and not flag_name.startswith("--no-"):
-                    flags.append("--no-" + flag_name[2:])
-            results[path] = sorted(set(flags))
-
-    if prefix:
-        results[prefix] = sorted(subcommand_names)
-
-    return results
-
-
-def get_gcloud_flags_from_help(cmd_path: str) -> list[str]:
-    """Parse `gcloud <cmd> --help` output to extract flags.
-
-    The static completion tree is occasionally incomplete for specific
-    commands; --help is the authoritative source of user-facing flags.
-    """
-    result = subprocess.run(
-        ["gcloud"] + cmd_path.split() + ["--help"],
-        capture_output=True,
-        text=True,
-        timeout=30,
-    )
-    output = result.stdout + result.stderr
-    flags = set()
-    for match in re.finditer(r"(--[a-z][a-z0-9-]*)", output):
-        flags.add(match.group(1))
-    return sorted(flags)
-
-
-def probe_gcloud_command(cmd_path: str) -> list[str] | None:
-    """Run `gcloud <cmd_path> --help` and return the parsed flags if the
-    command exists, otherwise None.
-
-    Used to backfill commands missing from the SDK's static completion
-    tree — gcloud hides some real subcommands such as
-    `gcloud logging cmek-settings update`.
-    """
-    try:
-        result = subprocess.run(
-            ["gcloud"] + cmd_path.split() + ["--help"],
-            capture_output=True,
-            text=True,
-            timeout=30,
-        )
-    except (subprocess.TimeoutExpired, FileNotFoundError):
-        return None
-    if result.returncode != 0:
-        return None
-    output = result.stdout + result.stderr
-    flags = set()
-    for match in re.finditer(r"(--[a-z][a-z0-9-]*)", output):
-        flags.add(match.group(1))
-    return sorted(flags)
-
-
-_GCLOUD_SUBCOMMAND_RE = re.compile(r"^[a-z][a-z0-9-]*$")
-
-
-def _iter_gcloud_policy_invocations() -> Iterator[list[str]]:
-    """Yield the post-``gcloud`` token list for every gcloud invocation in an
-    ``id: cli`` remediation in the GCP policy.
-
-    Centralises the YAML + fenced-block + line-continuation parsing so the
-    per-token filtering logic stays in the callers.
-    """
-    if not GCLOUD_POLICY_FILE.exists():
-        return
-
-    content = GCLOUD_POLICY_FILE.read_text()
-    for match in re.finditer(
-        r"- id: cli\s*\n\s+desc: \|\s*\n(.*?)(?=\n\s+- id: |\n\s+refs:|\n  - uid: |\Z)",
-        content,
-        re.DOTALL,
-    ):
-        for fence in re.finditer(
-            r"```bash\s*\n(.*?)```", match.group(1), re.DOTALL
-        ):
-            joined = re.sub(r"\\\s*\n\s*", " ", fence.group(1))
-            for line in joined.split("\n"):
-                line = line.strip()
-                if not line.startswith("gcloud "):
-                    continue
-                yield line.split()[1:]
-
-
-def detect_gcloud_policy_command_paths() -> set[str]:
-    """Return the bare gcloud command paths referenced by `id: cli` policy
-    remediations.
-
-    A path is the longest leading run of lowercase-hyphen tokens before any
-    flag or positional placeholder (`INSTANCE_NAME`, `[ZONE]`, `<KEY>`).
-    The result is suitable for direct lookup against the in-memory
-    commands DB.
-    """
-    paths: set[str] = set()
-    for tokens in _iter_gcloud_policy_invocations():
-        cmd_parts: list[str] = []
-        for p in tokens:
-            if p.startswith("-") or not _GCLOUD_SUBCOMMAND_RE.match(p):
-                break
-            cmd_parts.append(p)
-        if cmd_parts:
-            paths.add(" ".join(cmd_parts))
-    return paths
-
-
-def detect_gcloud_policy_commands(commands_db: dict[str, list[str]]) -> set[str]:
-    """Return the set of known gcloud command paths used in the GCP policy."""
-    policy_commands: set[str] = set()
-    for tokens in _iter_gcloud_policy_invocations():
-        cmd_parts: list[str] = []
-        for p in tokens:
-            if p.startswith("-"):
-                break
-            cmd_parts.append(p)
-        candidate = " ".join(cmd_parts)
-        if candidate in commands_db:
-            policy_commands.add(candidate)
-    return policy_commands
-
-
-def build_gcloud_commands_db() -> dict[str, list[str]]:
-    """Build an in-memory gcloud commands database by walking the SDK's
-    static completion tree, scoped to services used in the GCP policy, and
-    supplementing policy-referenced leaf commands with flags parsed from
-    `gcloud --help`.
-
-    Returns an empty dict if the policy file has no gcloud commands or the
-    gcloud SDK is not installed.
-    """
-    services = detect_gcloud_services_from_policy()
-    if not services:
-        return {}
-
-    completions_dir = find_gcloud_completions_path()
-    if not completions_dir:
-        print(
-            "Error: gcloud CLI not found in PATH.\n"
-            "\n"
-            "Validating gcloud remediation commands requires the Google Cloud\n"
-            "SDK to be installed locally — the static completion tree bundled\n"
-            "with the SDK is the source of truth for valid commands and flags.\n"
-            "\n"
-            "Install options:\n"
-            "  macOS (Homebrew): brew install --cask google-cloud-sdk\n"
-            "  Debian/Ubuntu:    https://cloud.google.com/sdk/docs/install-sdk#deb\n"
-            "  RHEL/Fedora:      https://cloud.google.com/sdk/docs/install-sdk#rpm\n"
-            "  Other:            https://cloud.google.com/sdk/docs/install\n"
-            "\n"
-            "After installing, ensure `gcloud` is on your PATH and re-run:\n"
-            f"  python3 {sys.argv[0]} gcp\n"
-            "\n"
-            "To skip gcloud validation, run a non-gcp target instead\n"
-            "(e.g. `aws`, `azure`, or `oci`).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    sys.path.insert(0, completions_dir)
-    try:
-        from gcloud_completions import STATIC_COMPLETION_CLI_TREE as tree
-    except ImportError as e:
-        print(
-            f"Error: Found gcloud at '{completions_dir}' but could not import\n"
-            f"gcloud_completions: {e}\n"
-            "\n"
-            "This usually indicates a broken or partial SDK install. Try\n"
-            "reinstalling the Google Cloud SDK (see\n"
-            "https://cloud.google.com/sdk/docs/install).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    global_flags = sorted(tree.get("flags", {}).keys())
-
-    commands: dict[str, list[str]] = {}
-    all_commands = tree.get("commands", {})
-    for svc in services:
-        svc_node = all_commands.get(svc)
-        if svc_node is None:
-            print(
-                f"Warning: gcloud service '{svc}' not found in SDK",
-                file=sys.stderr,
-            )
-            continue
-        commands.update(walk_gcloud_tree({"commands": {svc: svc_node}}, ""))
-
-    for key, value in commands.items():
-        if value and isinstance(value[0], str) and value[0].startswith("-"):
-            commands[key] = sorted(set(value + global_flags))
-
-    policy_commands = detect_gcloud_policy_commands(commands)
-    if policy_commands:
-        def _merge(cmd_name: str) -> tuple[str, list[str]]:
-            try:
-                help_flags = get_gcloud_flags_from_help(cmd_name)
-            except Exception:
-                help_flags = []
-            existing = commands.get(cmd_name, [])
-            return cmd_name, sorted(set(existing + help_flags))
-
-        with concurrent.futures.ThreadPoolExecutor(max_workers=8) as pool:
-            for cmd_name, merged in pool.map(_merge, policy_commands):
-                commands[cmd_name] = merged
-
-    # Backfill policy-referenced commands that aren't in the static
-    # completion tree (e.g. `gcloud logging cmek-settings update`) by
-    # probing `gcloud <cmd> --help` and using its exit code as ground truth.
-    policy_paths = detect_gcloud_policy_command_paths()
-    missing_paths = sorted(p for p in policy_paths if p not in commands)
-    if missing_paths:
-        with concurrent.futures.ThreadPoolExecutor(max_workers=8) as pool:
-            for cmd_path, flags in zip(
-                missing_paths, pool.map(probe_gcloud_command, missing_paths)
-            ):
-                if flags is None:
-                    continue
-                commands[cmd_path] = sorted(set(flags + global_flags))
-
-    return commands
-
-
-def parse_gcloud_command(cmd: str, commands_db: dict[str, list[str]]) -> tuple[str, list[str]]:
-    """Parse a gcloud command into (command_path, flags).
-
-    gcloud commands have variable-depth paths (e.g. 'compute instances create').
-    We match the longest known command path from the database. If no match is
-    found, the raw command path is returned so the caller can report it as unknown.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 2 or tokens[0] != "gcloud":
-        return "", []
-
-    # Find the longest matching command path by consuming tokens until
-    # we hit a flag or run out of matching commands
-    parts = tokens[1:]  # skip 'gcloud'
-    command_path = ""
-    raw_path_parts = []
-    for i in range(len(parts)):
-        if parts[i].startswith("-"):
-            break
-        raw_path_parts.append(parts[i])
-        candidate = " ".join(parts[: i + 1])
-        if candidate in commands_db:
-            command_path = candidate
-
-    # If no match found, return the raw path so the caller can report it
-    if not command_path and raw_path_parts:
-        command_path = " ".join(raw_path_parts)
-
-    # Extract flags from remaining tokens
-    flags = []
-    for token in parts:
-        if token.startswith("--"):
-            flag = token.split("=")[0]
-            flags.append(flag)
-
-    return command_path, flags
-
-
-def validate_gcloud_command(
-    command_path: str,
-    flags: list[str],
-    commands_db: dict[str, list[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed gcloud command against the commands database."""
-    errors = []
-
-    if command_path not in commands_db:
-        errors.append(f"unknown command 'gcloud {command_path}'")
-        return False, errors
-
-    valid_flags = set(commands_db[command_path])
-    for flag in flags:
-        if flag not in valid_flags:
-            errors.append(f"unknown flag '{flag}' for 'gcloud {command_path}'")
-
-    return len(errors) == 0, errors
-
-
-def validate_gcloud() -> tuple[int, int]:
-    """Validate gcloud CLI commands. Returns (pass_count, fail_count)."""
-    if not GCLOUD_POLICY_FILE.exists():
-        print(f"Error: Policy file not found: {GCLOUD_POLICY_FILE}", file=sys.stderr)
-        sys.exit(1)
-
-    commands_db = build_gcloud_commands_db()
-    if not commands_db:
-        return 0, 0
-
-    content = GCLOUD_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-
-    pass_count = 0
-    fail_count = 0
-
-    policy_relpath = str(GCLOUD_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "gcloud", block_line)
-        for cmd, line_num in commands:
-            command_path, flags = parse_gcloud_command(cmd, commands_db)
-
-            if not command_path:
-                continue
-
-            is_valid, errors = validate_gcloud_command(
-                command_path, flags, commands_db
-            )
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": truncate_cmd(cmd),
-                    "errors": errors,
-                    "cloud": "gcp",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# doctl (DigitalOcean) validation
-# ---------------------------------------------------------------------------
-#
-# Unlike aws/oci/gcp, doctl is a Go binary with no Python introspection
-# surface and no static completion tree we can parse. We walk `doctl --help`
-# breadth-first, parsing each subcommand and flag list out of the Cobra-
-# rendered help text. The walk is parallelized; a full traversal of the
-# ~475-command tree completes in roughly one second.
-
-DO_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-digitalocean-security.mql.yaml"
-
-# Cobra-rendered section headers like "Available Commands:", "Manage
-# DigitalOcean Resources:", "View Billing:". The top-level `doctl --help`
-# uses category headers instead of "Available Commands:".
-_DOCTL_SECTION_HEADER = re.compile(r"^[A-Z][^:\n]*:\s*$")
-
-# Subcommand line: "  <name>   <description>" — two-space indent, then a
-# lowercase identifier, then whitespace before the description. Cobra pads
-# names to the longest one plus a single space, so the longest subcommand in
-# a group has exactly one space before its description.
-_DOCTL_SUBCOMMAND_LINE = re.compile(r"^  ([a-z][a-z0-9-]*)\s+\S")
-
-# Flag line: "      --flag-name <type>   <description>" or
-# "  -X, --flag-name <type>   <description>". We only capture the long form.
-_DOCTL_FLAG_LINE = re.compile(r"^\s+(?:-\w,\s+)?(--[a-z][a-z0-9-]*)")
-
-
-def _doctl_help(path: str) -> str:
-    """Return stdout+stderr of `doctl <path> --help`."""
-    args = ["doctl"] + (path.split() if path else []) + ["--help"]
-    r = subprocess.run(args, capture_output=True, text=True, timeout=15)
-    return r.stdout + r.stderr
-
-
-def _parse_doctl_help(text: str) -> tuple[list[str], list[str]]:
-    """Extract (subcommands, flags) from a doctl --help blob.
-
-    Handles both the top-level layout (category headers like "Manage
-    DigitalOcean Resources:") and group/leaf layouts ("Available Commands:"
-    / "Flags:" / "Global Flags:").
-    """
-    subcommands: list[str] = []
-    flags: list[str] = []
-    in_flags = False
-    in_cmds = False
-
-    for line in text.split("\n"):
-        if line.startswith("Flags:") or line.startswith("Global Flags:"):
-            in_flags = True
-            in_cmds = False
-            continue
-        if _DOCTL_SECTION_HEADER.match(line):
-            in_flags = False
-            # Example invocations are indented like subcommand lines but
-            # start with "doctl"; don't mine the Examples/Usage sections.
-            in_cmds = not line.startswith(("Examples:", "Usage:"))
-            continue
-        if line.strip() == "":
-            continue
-
-        if in_flags:
-            m = _DOCTL_FLAG_LINE.match(line)
-            if m:
-                flags.append(m.group(1))
-        elif in_cmds:
-            m = _DOCTL_SUBCOMMAND_LINE.match(line)
-            if m:
-                subcommands.append(m.group(1))
-
-    return sorted(set(subcommands)), sorted(set(flags))
-
-
-def detect_doctl_services_from_policy() -> list[str]:
-    """Return the set of top-level doctl services used in the DO policy."""
-    if not DO_POLICY_FILE.exists():
-        return []
-
-    content = DO_POLICY_FILE.read_text()
-    services = set()
-    for match in re.finditer(r"```bash\s*\n(.*?)```", content, re.DOTALL):
-        block = match.group(1)
-        joined = re.sub(r"\\\s*\n\s*", " ", block)
-        for line in joined.split("\n"):
-            line = line.strip()
-            if line.startswith("doctl "):
-                parts = line.split()
-                if len(parts) >= 2 and not parts[1].startswith("-"):
-                    services.add(parts[1])
-    return sorted(services)
-
-
-def build_doctl_commands_db() -> dict[str, list[str]]:
-    """Build an in-memory doctl commands database by walking the Cobra help
-    tree breadth-first, scoped to services used in the DO policy.
-
-    Returns an empty dict if the policy file has no doctl commands. Exits
-    with a helpful error if doctl is not installed.
-    """
-    services = detect_doctl_services_from_policy()
-    if not services:
-        return {}
-
-    doctl_path = subprocess.run(
-        ["which", "doctl"], capture_output=True, text=True
-    ).stdout.strip()
-    if not doctl_path:
-        print(
-            "Error: doctl CLI not found in PATH.\n"
-            "\n"
-            "Validating doctl remediation commands requires the doctl CLI to\n"
-            "be installed locally — its `--help` output is the source of truth\n"
-            "for valid commands and flags.\n"
-            "\n"
-            "Install options:\n"
-            "  macOS (Homebrew): brew install doctl\n"
-            "  Linux (snap):     snap install doctl\n"
-            "  Other:            https://docs.digitalocean.com/reference/doctl/how-to/install/\n"
-            "\n"
-            "After installing, ensure `doctl` is on your PATH and re-run:\n"
-            f"  python3 {sys.argv[0]} digitalocean\n"
-            "\n"
-            "To skip doctl validation, run a different target instead\n"
-            "(e.g. `aws`, `azure`, `oci`, or `gcp`).",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    # Breadth-first walk, parallelizing the subprocess calls in each round.
-    # The executor is created once and reused across rounds so we don't pay
-    # thread-pool startup cost on every level of the tree.
-    visited: set[str] = set()
-    results: dict[str, dict] = {}
-    queue = list(services)
-
-    with concurrent.futures.ThreadPoolExecutor(max_workers=16) as pool:
-        while queue:
-            future_to_path = {}
-            for path in queue:
-                if path in visited:
-                    continue
-                visited.add(path)
-                future_to_path[pool.submit(_doctl_help, path)] = path
-            next_queue: list[str] = []
-            for fut, path in future_to_path.items():
-                try:
-                    text = fut.result()
-                except subprocess.TimeoutExpired:
-                    print(
-                        f"Warning: timed out fetching `doctl {path} --help`",
-                        file=sys.stderr,
-                    )
-                    continue
-                subs, flags = _parse_doctl_help(text)
-                results[path] = {"subcommands": subs, "flags": flags}
-                for s in subs:
-                    sub_path = f"{path} {s}".strip()
-                    if sub_path not in visited:
-                        next_queue.append(sub_path)
-            queue = next_queue
-
-    # Flatten {"compute firewall": {subcommands, flags}, ...} into the
-    # validator's expected shape: a flat map where group entries hold the
-    # list of valid subcommand names and leaf entries hold the flag list.
-    commands: dict[str, list[str]] = {}
-    for path, node in results.items():
-        if node["subcommands"]:
-            commands[path] = node["subcommands"]
-        else:
-            commands[path] = node["flags"]
-    return commands
-
-
-def parse_doctl_command(cmd: str, commands_db: dict[str, list[str]]) -> tuple[str, list[str]]:
-    """Parse a doctl command into (command_path, flags).
-
-    doctl commands have variable-depth paths (e.g. 'compute firewall add-rules').
-    We match the longest known command path from the database. If no match is
-    found, the raw command path (tokens before the first flag) is returned so
-    the caller can report it as unknown.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 2 or tokens[0] != "doctl":
-        return "", []
-
-    parts = tokens[1:]  # skip 'doctl'
-    command_path = ""
-    raw_path_parts = []
-    for i in range(len(parts)):
-        if parts[i].startswith("-"):
-            break
-        raw_path_parts.append(parts[i])
-        candidate = " ".join(parts[: i + 1])
-        if candidate in commands_db:
-            command_path = candidate
-
-    if not command_path and raw_path_parts:
-        command_path = " ".join(raw_path_parts)
-
-    flags = []
-    for token in parts:
-        if token.startswith("--"):
-            flags.append(token.split("=")[0])
-
-    return command_path, flags
-
-
-def validate_doctl_command(
-    command_path: str,
-    flags: list[str],
-    commands_db: dict[str, list[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed doctl command against the commands database."""
-    errors = []
-
-    if command_path not in commands_db:
-        errors.append(f"unknown command 'doctl {command_path}'")
-        return False, errors
-
-    entry = commands_db[command_path]
-    # Group entries hold subcommand names; leaf entries hold flag names
-    # (which start with `--`). If we landed on a group, the user is missing
-    # a required subcommand.
-    if entry and not entry[0].startswith("-"):
-        errors.append(
-            f"'doctl {command_path}' is a command group; missing subcommand"
-        )
-        return False, errors
-
-    valid_flags = set(entry)
-    for flag in flags:
-        if flag not in valid_flags:
-            errors.append(f"unknown flag '{flag}' for 'doctl {command_path}'")
-
-    return len(errors) == 0, errors
-
-
-def validate_digitalocean() -> tuple[int, int]:
-    """Validate doctl CLI commands. Returns (pass_count, fail_count)."""
-    if not DO_POLICY_FILE.exists():
-        print(f"Error: Policy file not found: {DO_POLICY_FILE}", file=sys.stderr)
-        sys.exit(1)
-
-    commands_db = build_doctl_commands_db()
-    if not commands_db:
-        return 0, 0
-
-    content = DO_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-
-    pass_count = 0
-    fail_count = 0
-
-    policy_relpath = str(DO_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "doctl", block_line)
-        for cmd, line_num in commands:
-            command_path, flags = parse_doctl_command(cmd, commands_db)
-
-            if not command_path:
-                continue
-
-            is_valid, errors = validate_doctl_command(
-                command_path, flags, commands_db
-            )
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": truncate_cmd(cmd),
-                    "errors": errors,
-                    "cloud": "digitalocean",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# Cloudflare API validation
-# ---------------------------------------------------------------------------
-#
-# Cloudflare's `flarectl` is read-only for zone settings and `wrangler` does
-# not cover the surface area this policy targets (SSL/TLS, WAF, security
-# level, 2FA enforcement, etc.). The realistic CLI path is `curl` against
-# the Cloudflare REST API, so the validator scans for those calls and
-# verifies each path + HTTP method — and, where the call sends a JSON
-# payload, the request body — against the official OpenAPI spec published
-# at https://github.com/cloudflare/api-schemas.
-#
-# Body validation covers the subset of JSON Schema the Cloudflare spec
-# actually uses for the endpoints in this policy: $ref, allOf, oneOf/anyOf,
-# type, enum, properties/required, and array items. Angle-bracket
-# placeholders (`"<account-name>"`) act as wildcards that satisfy any
-# schema. For the generic `/zones/{zone_id}/settings/{setting_id}` PATCH —
-# whose spec body is a oneOf over every zone setting — the validator
-# resolves the per-setting component (`zones_<setting_id>_value`) from the
-# literal setting id in the URL, so wrong enum values and misspelled
-# payload fields fail instead of slipping through the union.
-#
-# The spec is large (~9 MB) and the upstream repo doesn't ship releases —
-# `main` is the only branch and it's auto-updated. We pin a known-good
-# commit SHA so validation is deterministic; bumping CLOUDFLARE_OPENAPI_SHA
-# is a deliberate maintainer action, like bumping doctl or the AWS CLI
-# version. The fetched spec is cached on disk under ~/.cache so we don't
-# re-download on every run.
-
-CLOUDFLARE_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-cloudflare-security.mql.yaml"
-
-CLOUDFLARE_OPENAPI_SHA = "a0e7cfa11b0d08b05a0b373f47ea722bd48ca7c4"
-CLOUDFLARE_OPENAPI_URL = (
-    f"https://raw.githubusercontent.com/cloudflare/api-schemas/"
-    f"{CLOUDFLARE_OPENAPI_SHA}/openapi.json"
-)
-CLOUDFLARE_API_PREFIX = "https://api.cloudflare.com/client/v4"
-
-
-def _cloudflare_cache_path() -> Path:
-    cache_dir = Path.home() / ".cache" / "cnspec-validation"
-    cache_dir.mkdir(parents=True, exist_ok=True)
-    return cache_dir / f"cloudflare-openapi-{CLOUDFLARE_OPENAPI_SHA[:12]}.json"
-
-
-def _load_cloudflare_openapi() -> dict:
-    """Return the parsed Cloudflare OpenAPI spec, downloading and caching
-    on first use."""
-    cache = _cloudflare_cache_path()
-    if not cache.exists():
-        print(
-            f"Fetching Cloudflare OpenAPI spec (sha {CLOUDFLARE_OPENAPI_SHA[:12]})...",
-            file=sys.stderr,
-        )
-        try:
-            with urllib.request.urlopen(CLOUDFLARE_OPENAPI_URL, timeout=60) as r:
-                cache.write_bytes(r.read())
-        except (urllib.error.URLError, TimeoutError) as e:
-            print(
-                f"Error: failed to download Cloudflare OpenAPI spec from\n"
-                f"  {CLOUDFLARE_OPENAPI_URL}\n"
-                f"  ({e})\n"
-                "\n"
-                "If you are behind a proxy or air-gapped, manually download the\n"
-                f"spec and save it to:\n  {cache}",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-    try:
-        return json.loads(cache.read_text())
-    except json.JSONDecodeError as e:
-        print(
-            f"Error: cached Cloudflare OpenAPI spec is corrupted\n"
-            f"  ({cache}: {e})\n"
-            "\n"
-            "Delete the cache file and re-run to re-download:\n"
-            f"  rm {cache}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-
-def _index_cloudflare_paths(spec: dict) -> dict[str, set[str]]:
-    """Index OpenAPI paths so we can match a concrete URL path to a path
-    template and verify the HTTP method.
-
-    Returns {path_template: {http_method, ...}} where http_method is the
-    upper-cased verb. Path templates are stored verbatim from the spec
-    (e.g. "/zones/{zone_id}/settings/{setting_id}").
-    """
-    out: dict[str, set[str]] = {}
-    for path, ops in spec.get("paths", {}).items():
-        methods = {m.upper() for m in ops if m in (
-            "get", "post", "put", "patch", "delete", "head", "options",
-        )}
-        if methods:
-            out[path] = methods
-    return out
-
-
-# Placeholder tokens used in remediation examples (e.g. <zone-id>,
-# <account-id>). The validator treats these as wildcards that match any
-# OpenAPI path parameter ({zone_id}, {account_id}, ...).
-_CF_PLACEHOLDER_RE = re.compile(r"^<[a-z][a-z0-9-]*>$")
-
-
-def _match_cloudflare_path(curl_path: str, openapi_paths: dict[str, set[str]]) -> str | None:
-    """Find an OpenAPI path template matching a concrete curl URL path.
-
-    A segment matches if:
-      - both segments are literally equal, or
-      - the OpenAPI segment is a `{param}` template and the curl segment
-        is either a placeholder like `<zone-id>` or a concrete literal.
-
-    When multiple templates match, prefers the one with the most literal
-    (non-`{param}`) segments — so e.g. `/zones/{id}/dns_records/import`
-    wins over `/zones/{id}/dns_records/{record_id}` for a curl call to
-    `/zones/<zone-id>/dns_records/import`. Returns the matched path
-    template, or None.
-    """
-    curl_parts = curl_path.split("/")
-
-    # Exact-literal match first.
-    if curl_path in openapi_paths:
-        return curl_path
-
-    best: str | None = None
-    best_specificity = -1
-    for tmpl in openapi_paths:
-        tmpl_parts = tmpl.split("/")
-        if len(tmpl_parts) != len(curl_parts):
-            continue
-        if not all(_segment_matches(t, c) for t, c in zip(tmpl_parts, curl_parts)):
-            continue
-        specificity = sum(
-            1 for p in tmpl_parts if p and not (p.startswith("{") and p.endswith("}"))
-        )
-        if specificity > best_specificity:
-            best = tmpl
-            best_specificity = specificity
-    return best
-
-
-def _segment_matches(tmpl_seg: str, curl_seg: str) -> bool:
-    if tmpl_seg == curl_seg:
-        return True
-    if tmpl_seg.startswith("{") and tmpl_seg.endswith("}"):
-        # Templated segment matches any concrete value or angle-bracket
-        # placeholder in the curl URL.
-        return curl_seg != "" and "/" not in curl_seg
-    return False
-
-
-# Match a curl invocation. The URL may be quoted or unquoted; -X / --request
-# may appear before or after it.
-_CF_CURL_URL_RE = re.compile(
-    r'https?://api\.cloudflare\.com/client/v4(/[^\s\'"]+)'
-)
-_CF_CURL_METHOD_RE = re.compile(
-    r'(?:^|\s)(?:-X|--request)(?:\s+|=)([A-Z]+)'
-)
-# Any curl flag that supplies a request body. split_commands() has already
-# shlex-tokenized and re-joined the command, so the payload appears after
-# the flag with its shell quoting stripped but its JSON quoting intact.
-_CF_CURL_DATA_RE = re.compile(
-    r"(?:^|\s)(?:--data(?:-raw|-binary|-ascii)?|--json|-d)(?:\s+|=)"
-)
-
-# A whole-string placeholder value like "<account-name>". During body
-# validation these act as wildcards that satisfy any schema.
-_CF_PLACEHOLDER_VALUE_RE = re.compile(r"^<[A-Za-z0-9_.-]+>$")
-
-# Quote a bare (unquoted) placeholder so the payload parses as JSON, e.g.
-# {"max_age": <seconds>} -> {"max_age": "<seconds>"}. Placeholders already
-# inside JSON strings are excluded by the quote look-around.
-_CF_BARE_PLACEHOLDER_RE = re.compile(r'(?<![\w"<])<([A-Za-z0-9_.-]+)>(?![\w">])')
-
-# Known divergences between Cloudflare's OpenAPI spec and the documented
-# behavior of the API. Maps (METHOD, path_template) to the validation step
-# to skip:
-#   "body"     — skip request-body validation entirely.
-#   "required" — validate the body but skip required-property checks.
-CLOUDFLARE_BODY_EXEMPTIONS = {
-    # The token-roll endpoint declares its body as `type: object`, but the
-    # Cloudflare API docs (and the live API) require the literal JSON
-    # string "" as the request body.
-    ("PUT", "/user/tokens/{token_id}/value"): "body",
-    # Account update reuses the shared account schema, which marks `id` and
-    # `type` required because responses always carry them; the documented
-    # update payload sends only `name` (plus optional `settings`).
-    ("PUT", "/accounts/{account_id}"): "required",
-}
-
-
-def parse_cloudflare_curl(cmd: str) -> tuple[str, str, str | None] | None:
-    """Extract (HTTP_METHOD, URL_PATH, BODY) from a curl Cloudflare API call.
-
-    Returns None if the command is not a Cloudflare API curl. The method
-    defaults to GET when -X / --request is absent (matches curl's default
-    when no body is supplied; for POST/PATCH/PUT we always require an
-    explicit -X in remediation snippets). BODY is the raw --data payload,
-    or None when the command sends no body.
-    """
-    if "api.cloudflare.com/client/v4" not in cmd:
-        return None
-    if not cmd.lstrip().startswith("curl"):
-        return None
-
-    url_match = _CF_CURL_URL_RE.search(cmd)
-    if not url_match:
-        return None
-    path = url_match.group(1)
-    # Drop a fragment or query string if present.
-    path = path.split("?", 1)[0].split("#", 1)[0]
-
-    method_match = _CF_CURL_METHOD_RE.search(cmd)
-    method = method_match.group(1).upper() if method_match else "GET"
-
-    return method, path, _extract_curl_body(cmd)
-
-
-def _extract_curl_body(cmd: str) -> str | None:
-    """Return the payload passed via --data/-d/--json, or None.
-
-    The command string has been shlex-tokenized and space-joined, so the
-    payload's shell quotes are gone. A JSON object/array payload may
-    therefore contain spaces; scan to its balanced closing brace instead of
-    splitting on whitespace.
-    """
-    m = _CF_CURL_DATA_RE.search(cmd)
-    if not m:
-        return None
-    rest = cmd[m.end():]
-    if not rest:
-        return None
-    if rest[0] in "{[":
-        close = {"{": "}", "[": "]"}[rest[0]]
-        depth = 0
-        in_str = False
-        escaped = False
-        for i, ch in enumerate(rest):
-            if escaped:
-                escaped = False
-            elif ch == "\\":
-                escaped = True
-            elif ch == '"':
-                in_str = not in_str
-            elif not in_str:
-                if ch == rest[0]:
-                    depth += 1
-                elif ch == close:
-                    depth -= 1
-                    if depth == 0:
-                        return rest[: i + 1]
-        return rest  # unbalanced; json.loads will report the details
-    return rest.split(None, 1)[0]
-
-
-def _cf_resolve_ref(node, spec: dict):
-    """Follow a chain of $ref JSON pointers within the spec document."""
-    seen: set[str] = set()
-    while isinstance(node, dict) and "$ref" in node:
-        ref = node["$ref"]
-        if ref in seen:
-            return {}
-        seen.add(ref)
-        cur = spec
-        for part in ref.lstrip("#/").split("/"):
-            part = part.replace("~1", "/").replace("~0", "~")
-            cur = cur.get(part, {}) if isinstance(cur, dict) else {}
-        node = cur
-    return node
-
-
-def _cf_type_ok(value, t) -> bool:
-    if isinstance(t, list):
-        return any(_cf_type_ok(value, x) for x in t)
-    if t == "object":
-        return isinstance(value, dict)
-    if t == "array":
-        return isinstance(value, list)
-    if t == "string":
-        return isinstance(value, str)
-    if t == "boolean":
-        return isinstance(value, bool)
-    if t == "integer":
-        return isinstance(value, int) and not isinstance(value, bool)
-    if t == "number":
-        return isinstance(value, (int, float)) and not isinstance(value, bool)
-    if t == "null":
-        return value is None
-    return True
-
-
-def _cf_json_type_name(value) -> str:
-    if value is None:
-        return "null"
-    if isinstance(value, bool):
-        return "boolean"
-    if isinstance(value, int):
-        return "integer"
-    if isinstance(value, float):
-        return "number"
-    if isinstance(value, str):
-        return "string"
-    if isinstance(value, list):
-        return "array"
-    return "object"
-
-
-def _cf_validate_body(
-    value,
-    schema,
-    spec: dict,
-    loc: str,
-    errors: list[str],
-    check_required: bool = True,
-) -> None:
-    """Validate a parsed JSON value against a (subset of) JSON Schema.
-
-    Covers what the Cloudflare spec uses for the request bodies in this
-    policy: $ref, allOf, oneOf/anyOf, type, enum, object properties with
-    required/additionalProperties, and array items. Whole-string
-    placeholders pass any schema. Errors are appended to `errors` with
-    `loc` as the JSON-path-ish location prefix.
-    """
-    schema = _cf_resolve_ref(schema, spec)
-    if not isinstance(schema, dict) or not schema:
-        return
-    if isinstance(value, str) and _CF_PLACEHOLDER_VALUE_RE.match(value):
-        return
-
-    if "allOf" in schema:
-        # Shallow-merge the members into one object view so a property
-        # declared in one member isn't reported unknown by another.
-        merged = {k: v for k, v in schema.items() if k != "allOf"}
-        for member in schema["allOf"]:
-            member = _cf_resolve_ref(member, spec)
-            if not isinstance(member, dict):
-                continue
-            for key in ("type", "enum", "items", "additionalProperties"):
-                if key in member and key not in merged:
-                    merged[key] = member[key]
-            if "properties" in member:
-                merged.setdefault("properties", {}).update(member["properties"])
-            if "required" in member:
-                combined = merged.get("required", []) + member["required"]
-                merged["required"] = list(dict.fromkeys(combined))
-        _cf_validate_body(value, merged, spec, loc, errors, check_required)
-        return
-
-    variants = schema.get("oneOf") or schema.get("anyOf")
-    if variants:
-        for variant in variants:
-            variant_errors: list[str] = []
-            _cf_validate_body(
-                value, variant, spec, loc, variant_errors, check_required
-            )
-            if not variant_errors:
-                return
-        errors.append(f"{loc}: does not match any schema variant accepted here")
-        return
-
-    t = schema.get("type")
-    if t and not _cf_type_ok(value, t):
-        errors.append(f"{loc}: expected {t}, got {_cf_json_type_name(value)}")
-        return
-
-    if "enum" in schema and value not in schema["enum"]:
-        allowed = ", ".join(repr(x) for x in schema["enum"])
-        errors.append(f"{loc}: {value!r} is not one of [{allowed}]")
-        return
-
-    if isinstance(value, dict):
-        props = schema.get("properties")
-        if isinstance(props, dict):
-            additional = schema.get("additionalProperties")
-            # JSON Schema treats an absent additionalProperties as "allow
-            # anything"; we deliberately flag unknown keys anyway. The spec
-            # documents every accepted field for these endpoints, and a
-            # typo'd field name the API would silently ignore is exactly
-            # the bug this validator exists to catch. Only an explicit
-            # additionalProperties (true or a schema) relaxes the check.
-            if not (additional is True or isinstance(additional, dict)):
-                for key in value:
-                    if key not in props:
-                        errors.append(
-                            f"{loc}: unknown property '{key}' "
-                            f"(documented: {', '.join(sorted(props))})"
-                        )
-            for key, val in value.items():
-                if key in props:
-                    _cf_validate_body(
-                        val, props[key], spec, f"{loc}.{key}", errors, check_required
-                    )
-        if check_required:
-            for req in schema.get("required", []):
-                if req in value:
-                    continue
-                # readOnly properties are response-only; their presence in
-                # `required` does not apply to request bodies.
-                req_schema = {}
-                if isinstance(props, dict) and req in props:
-                    req_schema = _cf_resolve_ref(props[req], spec)
-                if isinstance(req_schema, dict) and req_schema.get("readOnly"):
-                    continue
-                errors.append(f"{loc}: missing required property '{req}'")
-    elif isinstance(value, list):
-        items = schema.get("items")
-        if items:
-            for i, val in enumerate(value):
-                _cf_validate_body(
-                    val, items, spec, f"{loc}[{i}]", errors, check_required
-                )
-
-
-def _cf_parse_body_json(body: str):
-    """Parse a curl payload as JSON. Returns (value, error_message)."""
-    try:
-        return json.loads(body), None
-    except json.JSONDecodeError:
-        pass
-    # Retry with bare placeholders quoted, e.g. {"max_age": <seconds>}.
-    try:
-        return json.loads(_CF_BARE_PLACEHOLDER_RE.sub(r'"<\1>"', body)), None
-    except json.JSONDecodeError as e:
-        return None, f"request body is not valid JSON ({e.msg})"
-
-
-def _cf_request_body_schema(spec: dict, tmpl: str, method: str):
-    """Return (schema, required) for an operation's JSON request body.
-
-    schema is None when the spec defines no request body for the operation.
-    """
-    op = spec.get("paths", {}).get(tmpl, {}).get(method.lower(), {})
-    request_body = _cf_resolve_ref(op.get("requestBody"), spec)
-    if not isinstance(request_body, dict) or not request_body:
-        return None, False
-    schema = (
-        request_body.get("content", {})
-        .get("application/json", {})
-        .get("schema")
-    )
-    return schema, bool(request_body.get("required"))
-
-
-def _cf_setting_request_schema(setting_id: str, components: dict) -> dict | None:
-    """Return the narrowed request-body schema for a zone setting id, or
-    None when the spec has no per-setting component (an unknown setting).
-
-    Most settings follow the `zones_<setting_id>_value` convention with a
-    {"value": ...} body; a few embed a service prefix in the component
-    name (`zones_cache-rules_aegis_value` for `settings/aegis`), and
-    ssl_recommender uses an `_enabled` component with an
-    {"enabled": ...} body instead.
-    """
-    prop = "value"
-    name = f"zones_{setting_id}_value"
-    if name not in components:
-        suffix = f"_{setting_id}_value"
-        matches = sorted(
-            n for n in components
-            if n.startswith("zones_") and n.endswith(suffix)
-        )
-        if matches:
-            # Unambiguous in the pinned spec; sorted() keeps the pick
-            # deterministic should a future spec bump introduce a collision.
-            name = matches[0]
-        elif f"zones_{setting_id}_enabled" in components:
-            prop = "enabled"
-            name = f"zones_{setting_id}_enabled"
-        else:
-            return None
-    return {
-        "type": "object",
-        "properties": {prop: {"$ref": f"#/components/schemas/{name}"}},
-        "required": [prop],
-    }
-
-
-def validate_cloudflare_curl(
-    method: str,
-    path: str,
-    body: str | None,
-    spec: dict,
-    openapi_paths: dict[str, set[str]],
-) -> tuple[bool, list[str]]:
-    """Validate a parsed Cloudflare curl call against the OpenAPI spec."""
-    errors: list[str] = []
-
-    matched = _match_cloudflare_path(path, openapi_paths)
-    if not matched:
-        errors.append(f"unknown Cloudflare API path '{path}'")
-        return False, errors
-
-    if method not in openapi_paths[matched]:
-        allowed = sorted(openapi_paths[matched])
-        errors.append(
-            f"method '{method}' not supported on '{matched}' "
-            f"(supported: {', '.join(allowed)})"
-        )
-        return False, errors
-
-    components = spec.get("components", {}).get("schemas", {})
-
-    # The settings path template matches any {setting_id}, so a misspelled
-    # setting id would otherwise pass path validation.
-    setting_schema = None
-    if matched == "/zones/{zone_id}/settings/{setting_id}":
-        last_segment = path.rsplit("/", 1)[-1]
-        if not _CF_PLACEHOLDER_RE.match(last_segment):
-            setting_schema = _cf_setting_request_schema(last_segment, components)
-            if setting_schema is None:
-                errors.append(f"unknown zone setting '{last_segment}'")
-                return False, errors
-
-    exemption = CLOUDFLARE_BODY_EXEMPTIONS.get((method, matched))
-    if exemption == "body":
-        return True, errors
-
-    schema, body_required = _cf_request_body_schema(spec, matched, method)
-
-    if body is None:
-        if schema is not None and body_required:
-            errors.append(
-                f"'{method} {matched}' requires a request body, "
-                "but the command sends none"
-            )
-            return False, errors
-        return True, errors
-
-    if body.startswith("@"):
-        # Payload is read from a file the snippet builds; nothing to inspect.
-        return True, errors
-
-    if schema is None:
-        errors.append(f"'{method} {matched}' does not accept a request body")
-        return False, errors
-
-    value, parse_error = _cf_parse_body_json(body)
-    if parse_error:
-        errors.append(parse_error)
-        return False, errors
-
-    # For the generic settings endpoint, narrow the oneOf-over-everything
-    # request schema down to the specific setting named in the URL so enum
-    # and field-name mistakes are actually caught.
-    if setting_schema is not None:
-        schema = setting_schema
-
-    _cf_validate_body(
-        value, schema, spec, "body", errors,
-        check_required=(exemption != "required"),
-    )
-    return (not errors), errors
-
-
-def validate_cloudflare() -> tuple[int, int]:
-    """Validate Cloudflare API curl commands. Returns (pass_count, fail_count)."""
-    if not CLOUDFLARE_POLICY_FILE.exists():
-        print(
-            f"Error: Policy file not found: {CLOUDFLARE_POLICY_FILE}",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    content = CLOUDFLARE_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-
-    # Skip the OpenAPI download entirely if there are no Cloudflare API
-    # curl calls in the policy.
-    has_cf_curl = any(
-        "api.cloudflare.com/client/v4" in b for b, _, _ in blocks
-    )
-    if not has_cf_curl:
-        return 0, 0
-
-    spec = _load_cloudflare_openapi()
-    openapi_paths = _index_cloudflare_paths(spec)
-
-    pass_count = 0
-    fail_count = 0
-
-    policy_relpath = str(CLOUDFLARE_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "curl", block_line)
-        for cmd, line_num in commands:
-            parsed = parse_cloudflare_curl(cmd)
-            if parsed is None:
-                continue
-            method, path, body = parsed
-
-            is_valid, errors = validate_cloudflare_curl(
-                method, path, body, spec, openapi_paths
-            )
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {method} {path}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {method} {path}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": f"{method} {path}",
-                    "errors": errors,
-                    "cloud": "cloudflare",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# Nutanix nCLI validation
-# ---------------------------------------------------------------------------
-#
-# ncli is a proprietary Java CLI that Nutanix only distributes from a running
-# Prism cluster, so unlike aws/gcloud/doctl it cannot be installed in CI and
-# introspected. Instead we validate against a checked-in grammar generated by
-# dump_ncli_commands.py from the AOS Command Reference on portal.nutanix.com
-# (see that script for the pinned book version and how to regenerate).
-
-NUTANIX_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-nutanix-security.mql.yaml"
-NCLI_COMMANDS_FILE = CMD_DATA_DIR / "ncli_commands.json"
-
-# Entity short forms ncli accepts at runtime but the Command Reference does
-# not document. Only add entries attested in Nutanix's own documentation or
-# KB articles.
-NCLI_ENTITY_ALIASES = {
-    # `ncli ngt list` is the canonical NGT status check in Nutanix KBs,
-    # e.g. https://portal.nutanix.com/kb/3567
-    "ngt": "nutanix-guest-tools",
-}
-
-
-def _load_ncli_db() -> dict:
-    """Load the generated ncli grammar; exits if it has not been generated."""
-    if not NCLI_COMMANDS_FILE.exists():
-        print(
-            f"Error: Commands database not found: {NCLI_COMMANDS_FILE}\n"
-            f"Run dump_ncli_commands.py first to generate it.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    return json.loads(NCLI_COMMANDS_FILE.read_text())["entities"]
-
-
-def parse_ncli_command(cmd: str) -> tuple[str, str, list[str], list[str]] | None:
-    """Parse an ncli command into (entity, operation, params, stray_tokens).
-
-    ncli commands have a fixed shape: `ncli <entity> <operation>` followed
-    by `param=value` assignments. Tokens after the operation that lack an
-    '=' are returned as stray tokens so the caller can flag them.
-    """
-    try:
-        tokens = shlex.split(cmd)
-    except ValueError:
-        tokens = cmd.split()
-
-    if len(tokens) < 3 or tokens[0] != "ncli":
-        return None
-
-    entity, operation = tokens[1], tokens[2]
-    params = []
-    stray = []
-    for token in tokens[3:]:
-        if "=" in token:
-            params.append(token.split("=", 1)[0])
-        else:
-            stray.append(token)
-    return entity, operation, params, stray
-
-
-def validate_ncli_command(
-    entity: str,
-    operation: str,
-    params: list[str],
-    stray: list[str],
-    entities_db: dict,
-) -> tuple[bool, list[str]]:
-    """Validate a parsed ncli command against the generated grammar."""
-    errors = []
-
-    canonical_entity = NCLI_ENTITY_ALIASES.get(entity, entity)
-    if canonical_entity not in entities_db:
-        errors.append(f"unknown entity 'ncli {entity}'")
-        return False, errors
-
-    operations = entities_db[canonical_entity]["operations"]
-    canonical_op = None
-    for op_name, op in operations.items():
-        if operation == op_name or operation in op["aliases"]:
-            canonical_op = op_name
-            break
-    if canonical_op is None:
-        errors.append(f"unknown operation '{operation}' for 'ncli {entity}'")
-        return False, errors
-
-    op = operations[canonical_op]
-    valid_params = set(op["required"]) | set(op["optional"])
-    for param in params:
-        if param not in valid_params:
-            errors.append(
-                f"unknown parameter '{param}' for 'ncli {entity} {operation}'"
-            )
-
-    # Required parameters must appear even in template commands: a
-    # remediation that omits one fails when run as written, and placeholder
-    # values like name=<directory-name> satisfy this check.
-    for req in op["required"]:
-        if req not in params:
-            errors.append(
-                f"missing required parameter '{req}' for "
-                f"'ncli {entity} {operation}'"
-            )
-
-    for token in stray:
-        errors.append(
-            f"unexpected token '{token}' for 'ncli {entity} {operation}' "
-            "(ncli arguments use param=value form)"
-        )
-
-    return len(errors) == 0, errors
-
-
-def validate_nutanix() -> tuple[int, int]:
-    """Validate Nutanix ncli commands. Returns (pass_count, fail_count)."""
-    entities_db = _load_ncli_db()
-
-    if not NUTANIX_POLICY_FILE.exists():
-        print(f"Error: Policy file not found: {NUTANIX_POLICY_FILE}", file=sys.stderr)
-        sys.exit(1)
-
-    content = NUTANIX_POLICY_FILE.read_text()
-    blocks = extract_bash_blocks(content)
-    policy_relpath = str(NUTANIX_POLICY_FILE.resolve().relative_to(Path.cwd()))
-
-    pass_count = 0
-    fail_count = 0
-
-    for block_text, block_line, uid in blocks:
-        commands = split_commands(block_text, "ncli", block_line)
-        for cmd, line_num in commands:
-            parsed = parse_ncli_command(cmd)
-            if parsed is None:
-                continue
-
-            is_valid, errors = validate_ncli_command(*parsed, entities_db)
-
-            if is_valid:
-                print(f"[PASS] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                pass_count += 1
-            else:
-                print(f"[FAIL] {uid}")
-                print(f"       {truncate_cmd(cmd)}")
-                for error in errors:
-                    print(f"       {error}")
-                fail_count += 1
-                FAILURES.append({
-                    "file": policy_relpath,
-                    "line": line_num,
-                    "uid": uid,
-                    "command": truncate_cmd(cmd),
-                    "errors": errors,
-                    "cloud": "nutanix",
-                })
-
-    return pass_count, fail_count
-
-
-# ---------------------------------------------------------------------------
-# GitHub Actions annotations
-# ---------------------------------------------------------------------------
-
-def emit_github_annotations() -> None:
-    """Print GitHub Actions workflow commands for each failure.
-
-    These produce inline annotations on the PR Files tab, regardless of
-    whether the annotated file is part of the PR diff.
-    See https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/workflow-commands-for-github-actions#setting-an-error-message
-    """
-    for r in FAILURES:
-        msg = "; ".join(r["errors"]) + f" — {r['command']}"
-        title = f"{r['cloud'].upper()} CLI validation ({r['uid']})"
-        # Workflow command special characters must be encoded
-        msg = msg.replace("%", "%25").replace("\r", "%0D").replace("\n", "%0A")
-        title = title.replace("%", "%25").replace("\r", "%0D").replace("\n", "%0A").replace(",", "%2C").replace("::", "%3A%3A")
-        print(f"::error file={r['file']},line={r['line']},title={title}::{msg}")
-
-
-# ---------------------------------------------------------------------------
-# Main
-# ---------------------------------------------------------------------------
-
-def main():
-    args = sys.argv[1:]
-    github_actions = False
-    target = "all"
-
-    # Parse flags
-    positional = []
-    for arg in args:
-        if arg == "--github-actions":
-            github_actions = True
-        else:
-            positional.append(arg)
-
-    if positional:
-        target = positional[0]
-
-    if target not in ("all", *VALIDATORS):
-        print(
-            f"Unknown validator: {target}\n"
-            f"Usage: {sys.argv[0]} [{'|'.join(['all'] + VALIDATORS)}] [--github-actions]",
-            file=sys.stderr,
-        )
-        sys.exit(2)
-
-    total_pass = 0
-    total_fail = 0
-
-    if target in ("all", "aws"):
-        p, f = validate_aws()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "azure"):
-        p, f = validate_azure()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "oci"):
-        p, f = validate_oci()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "gcp"):
-        p, f = validate_gcloud()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "digitalocean"):
-        p, f = validate_digitalocean()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "cloudflare"):
-        p, f = validate_cloudflare()
-        total_pass += p
-        total_fail += f
-
-    if target in ("all", "nutanix"):
-        p, f = validate_nutanix()
-        total_pass += p
-        total_fail += f
-
-    if github_actions:
-        emit_github_annotations()
-
-    print(f"\n{total_pass} passed, {total_fail} failed", file=sys.stderr)
-    sys.exit(1 if total_fail > 0 else 0)
-
+from validators.main import main
 
 if __name__ == "__main__":
     main()
