+++ content/validation/validators/__init__.py
@@ -1,0 +1,3 @@
+# Copyright Mondoo, Inc. 2024, 2026
+# SPDX-License-Identifier: BUSL-1.1
+"""Validators for CLI/API commands in policy remediation sections."""
