+++ content/validation/validators/nutanix.py
@@ -1,0 +1,170 @@
+# Copyright Mondoo, Inc. 2024, 2026
+# SPDX-License-Identifier: BUSL-1.1
+# Nutanix nCLI validation.
+
+import json
+import shlex
+import sys
+
+from pathlib import Path
+
+from .common import CMD_DATA_DIR, FAILURES, SCRIPT_DIR, extract_bash_blocks, policy_relpath, split_commands, truncate_cmd
+
+
+# ---------------------------------------------------------------------------
+# Nutanix nCLI validation
+# ---------------------------------------------------------------------------
+#
+# ncli is a proprietary Java CLI that Nutanix only distributes from a running
+# Prism cluster, so unlike aws/gcloud/doctl it cannot be installed in CI and
+# introspected. Instead we validate against a checked-in grammar generated by
+# dump_ncli_commands.py from the AOS Command Reference on portal.nutanix.com
+# (see that script for the pinned book version and how to regenerate).
+
+NUTANIX_POLICY_FILE = SCRIPT_DIR / ".." / "mondoo-nutanix-security.mql.yaml"
+NCLI_COMMANDS_FILE = CMD_DATA_DIR / "ncli_commands.json"
+
+# Entity short forms ncli accepts at runtime but the Command Reference does
+# not document. Only add entries attested in Nutanix's own documentation or
+# KB articles.
+NCLI_ENTITY_ALIASES = {
+    # `ncli ngt list` is the canonical NGT status check in Nutanix KBs,
+    # e.g. https://portal.nutanix.com/kb/3567
+    "ngt": "nutanix-guest-tools",
+}
+
+
+def _load_ncli_db() -> dict:
+    """Load the generated ncli grammar; exits if it has not been generated."""
+    if not NCLI_COMMANDS_FILE.exists():
+        print(
+            f"Error: Commands database not found: {NCLI_COMMANDS_FILE}\n"
+            f"Run dump_ncli_commands.py first to generate it.",
+            file=sys.stderr,
+        )
+        sys.exit(1)
+    return json.loads(NCLI_COMMANDS_FILE.read_text())["entities"]
+
+
+def parse_ncli_command(cmd: str) -> tuple[str, str, list[str], list[str]] | None:
+    """Parse an ncli command into (entity, operation, params, stray_tokens).
+
+    ncli commands have a fixed shape: `ncli <entity> <operation>` followed
+    by `param=value` assignments. Tokens after the operation that lack an
+    '=' are returned as stray tokens so the caller can flag them.
+    """
+    try:
+        tokens = shlex.split(cmd)
+    except ValueError:
+        tokens = cmd.split()
+
+    if len(tokens) < 3 or tokens[0] != "ncli":
+        return None
+
+    entity, operation = tokens[1], tokens[2]
+    params = []
+    stray = []
+    for token in tokens[3:]:
+        if "=" in token:
+            params.append(token.split("=", 1)[0])
+        else:
+            stray.append(token)
+    return entity, operation, params, stray
+
+
+def validate_ncli_command(
+    entity: str,
+    operation: str,
+    params: list[str],
+    stray: list[str],
+    entities_db: dict,
+) -> tuple[bool, list[str]]:
+    """Validate a parsed ncli command against the generated grammar."""
+    errors = []
+
+    canonical_entity = NCLI_ENTITY_ALIASES.get(entity, entity)
+    if canonical_entity not in entities_db:
+        errors.append(f"unknown entity 'ncli {entity}'")
+        return False, errors
+
+    operations = entities_db[canonical_entity]["operations"]
+    canonical_op = None
+    for op_name, op in operations.items():
+        if operation == op_name or operation in op["aliases"]:
+            canonical_op = op_name
+            break
+    if canonical_op is None:
+        errors.append(f"unknown operation '{operation}' for 'ncli {entity}'")
+        return False, errors
+
+    op = operations[canonical_op]
+    valid_params = set(op["required"]) | set(op["optional"])
+    for param in params:
+        if param not in valid_params:
+            errors.append(
+                f"unknown parameter '{param}' for 'ncli {entity} {operation}'"
+            )
+
+    # Required parameters must appear even in template commands: a
+    # remediation that omits one fails when run as written, and placeholder
+    # values like name=<directory-name> satisfy this check.
+    for req in op["required"]:
+        if req not in params:
+            errors.append(
+                f"missing required parameter '{req}' for "
+                f"'ncli {entity} {operation}'"
+            )
+
+    for token in stray:
+        errors.append(
+            f"unexpected token '{token}' for 'ncli {entity} {operation}' "
+            "(ncli arguments use param=value form)"
+        )
+
+    return len(errors) == 0, errors
+
+
+def validate_nutanix() -> tuple[int, int]:
+    """Validate Nutanix ncli commands. Returns (pass_count, fail_count)."""
+    entities_db = _load_ncli_db()
+
+    if not NUTANIX_POLICY_FILE.exists():
+        print(f"Error: Policy file not found: {NUTANIX_POLICY_FILE}", file=sys.stderr)
+        sys.exit(1)
+
+    content = NUTANIX_POLICY_FILE.read_text()
+    blocks = extract_bash_blocks(content)
+    relpath = policy_relpath(NUTANIX_POLICY_FILE)
+
+    pass_count = 0
+    fail_count = 0
+
+    for block_text, block_line, uid in blocks:
+        commands = split_commands(block_text, "ncli", block_line)
+        for cmd, line_num in commands:
+            parsed = parse_ncli_command(cmd)
+            if parsed is None:
+                continue
+
+            is_valid, errors = validate_ncli_command(*parsed, entities_db)
+
+            if is_valid:
+                print(f"[PASS] {uid}")
+                print(f"       {truncate_cmd(cmd)}")
+                pass_count += 1
+            else:
+                print(f"[FAIL] {uid}")
+                print(f"       {truncate_cmd(cmd)}")
+                for error in errors:
+                    print(f"       {error}")
+                fail_count += 1
+                FAILURES.append({
+                    "file": relpath,
+                    "line": line_num,
+                    "uid": uid,
+                    "command": truncate_cmd(cmd),
+                    "errors": errors,
+                    "cloud": "nutanix",
+                })
+
+    return pass_count, fail_count
